requirements.txt
websockets
wave
asyncio
Flask
flask-cors
gunicorn
firebase-admin
google-cloud-storage
cloudevents
google-cloud-speech
google-cloud-texttospeech